+++ NEWS.md
@@ -1,5 +1,6 @@
 # spant 2.19.0
 * Removed dependency on smoother package.
+* Added functions for generating BOLD and trapezoidal response shapes.
 
 # spant 2.18.0
 * Improved automated file type detection.